fix: ensure a single AsyncBus is created
