+++ lib/mfe-api-kit/src/request-api/core/asyncBus.ts
@@ -1,6 +1,7 @@
-import { ApiRegistrationEvent } from './apiRegistration';
 import { IEvoApi, IEvoApiParams, IEvoApiReturn, IEvoApiMethod } from './clientTypes';
 import { wait } from '../../utils';
+import { createEvoEventName } from './eventNamingStrategy';
+import { clientEvents } from './constants';
 
 const eventTimeout = Symbol('Registration wait error');
 const registrationDeclined = Symbol('Registration declined error');
@@ -19,10 +20,10 @@ export class AsyncBus<TApiClient extends IEvoApi> {
 
     constructor(public readonly clientName: string) {
         this.$id = `${clientName}-${Date.now() % 60000}`;
-        this.cleanUpRegistrationListener = ApiRegistrationEvent.createListener(this.clientRegistrationHandler);
+        this.cleanUpRegistrationListener = ApiRegisterEvent.createListener(this.clientRegistrationHandler);
     }
 
-    private clientRegistrationHandler = (e: ApiRegistrationEvent) => {
+    private clientRegistrationHandler = (e: ApiRegisterEvent) => {
         if (e.detail.clientName !== this.clientName) return;
 
         this.client = e.detail.client as TApiClient;
@@ -114,5 +115,69 @@ export class AsyncBus<TApiClient extends IEvoApi> {
 export class AsyncBusException extends Error {
     constructor(public readonly clientName: string, public readonly type: 'registrationTimeout' | 'registrationAborted') {
         super(`"${clientName}" registration failed`);
+    }
+}
+
+interface RegistrationEventDetails<TClient extends IEvoApi = IEvoApi> {
+    clientName: string;
+    client: TClient;
+}
+
+export class ApiRegisterEvent<TClient extends IEvoApi = IEvoApi> extends CustomEvent<RegistrationEventDetails<TClient>> {
+    static type = createEvoEventName(clientEvents.INIT_CLIENT);
+
+    constructor(clientName: string, client: TClient) {
+        super(ApiRegisterEvent.type, {
+            detail: {
+                clientName,
+                client,
+            },
+        });
+    }
+
+    static createListener(callback: (e: ApiRegisterEvent) => void) {
+        function onHandleCallback(e: Event) {
+            // If we use here "instanceof" event is handled incorrectly. I assume the reason is that module's code with declaration of ApiRegisterEvent is loaded multiple times.
+
+            callback(e as ApiRegisterEvent);
+        }
+
+        window.addEventListener(ApiRegisterEvent.type, onHandleCallback);
+        return () => window.removeEventListener(ApiRegisterEvent.type, onHandleCallback);
+    };
+
+    static dispatch(clientName: string, client: IEvoApi) {
+        console.log(`dispatch ${clientName}: ${ApiRegisterEvent.type}`);
+        window.dispatchEvent(new ApiRegisterEvent(clientName, client));
+    }
+}
+
+interface UnregisterEventDetails {
+    clientName: string;
+}
+
+export class ApiUnregisterEvent extends CustomEvent<UnregisterEventDetails> {
+    static type = createEvoEventName(clientEvents.INIT_CLIENT);
+
+    constructor(clientName: string) {
+        super(ApiUnregisterEvent.type, {
+            detail: {
+                clientName,
+            },
+        });
+    }
+
+    static createListener(callback: (e: ApiUnregisterEvent) => void) {
+        function onHandleCallback(e: Event) {
+            if (!(e instanceof ApiUnregisterEvent)) return;
+            callback(e);
+        }
+
+        window.addEventListener(ApiUnregisterEvent.type, onHandleCallback);
+        return () => window.removeEventListener(ApiUnregisterEvent.type, onHandleCallback);
+    }
+
+    static dispatch(clientName: string) {
+        window.dispatchEvent(new ApiUnregisterEvent(clientName));
     }
 }
